requirements.txt
pandas
ipykernel
langchain
langchain-groq
langchain_community
python-dotenv
pypdf
PyPDF2
chromadb
# faiss-gpu
sentence-transformers
huggingface_hub
langchain_chroma
langchain-groq
langchain_huggingface